src/main/java/ru/job4j/array/AlgoArray.java
package ru.job4j.array;

public class AlgoArray {

    public static int[] bubbleSort(int[] array) {
        int length = array.length;
        boolean swapped = false;
        for (int i = 0; i < length - 1; i++) {
            for (int j = i + 1; j < length; j++) {
                if (array[i] > array[j]) {
                    swap(array, i, j);
                    swapped = true;
                }
            }
            if (!swapped) {
                break;
            }
        }
        return array;
    }

    public static void swap(int[] array, int i, int j) {
        int temp = array[i];
        array[i] = array[j];
        array[j] = temp;
    }
}
